add download func; clean up code
