+++ app-shell/app/main/util.js
@@ -12,4 +12,20 @@ function waitUntilServerResponds (createWindow, windowUrl) {
     })
 }
 
-module.exports = {waitUntilServerResponds}
+function download (url, dest, cb) {
+  var file = fs.createWriteStream(dest);
+  var request = http.get(url, function(response) {
+    response.pipe(file);
+    file.on('finish', function() {
+      file.close(cb);
+    });
+  }).on('error', function(err) {
+    fs.unlink(dest);
+    if (cb) cb(err.message);
+  });
+};
+
+module.exports = {
+  download,
+  waitUntilServerResponds
+}
